fixed going out of scope issues ocurred on refactored code with sanity checks (msvc missed them, gcc didn't)
+ fixed descriptor add write set bufferInfo/imageInfo going out of scope
+ fixed depth attachment width and height for custom framebuffer being set too early for command buffers consumption
+ code cleanup

+++ src/Helpers/Descriptor.cpp
@@ -35,7 +35,7 @@ void DescriptorHelper::updateDescriptorSets() noexcept
 {
   m_deviceFuncs->vkUpdateDescriptorSets(
     m_device,
-    static_cast<uint32_t>(m_descWriteList.size()),
+    m_descWriteList.size(),
     m_descWriteList.data(),
     0,
     nullptr
